cli bookmark track: track default remote if --remote is omitted
This makes `jj bookmark track` consistent with `jj git push`, which
uses the default remote instead of all remotes if `--remote` is not
specified.
