+++ cli/src/commands/git/mod.rs
@@ -28,11 +28,7 @@ use clap::Subcommand;
 use clap::ValueEnum;
 use jj_lib::config::ConfigFile;
 use jj_lib::config::ConfigSource;
-use jj_lib::git;
-use jj_lib::git::UnexpectedGitBackendError;
-use jj_lib::ref_name::RemoteNameBuf;
 use jj_lib::ref_name::RemoteRefSymbol;
-use jj_lib::store::Store;
 
 use self::clone::GitCloneArgs;
 use self::clone::cmd_git_clone;
@@ -115,14 +111,6 @@ pub fn maybe_add_gitignore(workspace_command: &WorkspaceCommandHelper) -> Result
     } else {
         Ok(())
     }
-}
-
-fn get_single_remote(store: &Store) -> Result<Option<RemoteNameBuf>, UnexpectedGitBackendError> {
-    let mut names = git::get_all_remote_names(store)?;
-    Ok(match names.len() {
-        1 => names.pop(),
-        _ => None,
-    })
 }
 
 /// Sets repository level `trunk()` alias to the specified remote symbol.
